fix(data-table): single story folder and example update
