--- packages/carbon-web-components/examples/codesandbox/basic/components/data-table/defs.js
@@ -1,30 +1,0 @@
-/**
- * @license
- *
- * Copyright IBM Corp. 2022
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-/**
- * Table sort state.
- */
-const TABLE_SORT_DIRECTION = {
-  /**
-   * Not sorted.
-   */
-  NONE: 'none',
-
-  /**
-   * Sorted ascendingly.
-   */
-  ASCENDING: 'ascending',
-
-  /**
-   * Sorted descendingly.
-   */
-  DESCENDING: 'descending',
-};
-
-export default TABLE_SORT_DIRECTION;
