+++ packages/carbon-web-components/examples/codesandbox/basic/components/data-table/src/index.js
@@ -1,11 +1,10 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2020, 2022
+ * Copyright IBM Corp. 2020, 2023
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
 
 import '@carbon/web-components/es/components/data-table/index.js';
-import '@carbon/web-components/es/components/pagination/index.js';
