--- packages/carbon-web-components/examples/codesandbox/data-table/.babelrc
@@ -1,25 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "modules": false,
-        "targets": [
-          "last 1 version",
-          "Firefox ESR",
-          "not opera > 0",
-          "not op_mini > 0",
-          "not op_mob > 0",
-          "not android > 0",
-          "not edge > 0",
-          "not ie > 0",
-          "not ie_mob > 0"
-        ]
-      }
-    ]
-  ],
-  "plugins": [
-    ["@babel/plugin-proposal-decorators", { "decoratorsBeforeExport": true }],
-    ["@babel/plugin-transform-runtime", { "version": "7.3.0" }]
-  ]
-}
