--- packages/carbon-web-components/examples/codesandbox/data-table/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "carbon-web-components-getting-started",
-  "version": "0.1.0",
-  "private": true,
-  "description": "Sample project for getting started with the Web Components from the Carbon Design System.",
-  "license": "Apache-2",
-  "main": "index.html",
-  "scripts": {
-    "build": "parcel build index.html",
-    "start": "parcel index.html --no-hmr"
-  },
-  "dependencies": {
-    "@carbon/web-components": "latest",
-    "lit": "^2.6.0",
-    "lodash-es": "^4.17.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.10.0",
-    "@babel/plugin-proposal-decorators": "^7.10.0",
-    "@babel/plugin-transform-runtime": "^7.10.0",
-    "@types/lodash-es": "^4.17.0",
-    "parcel-bundler": "^1.10.0"
-  }
-}
