--- packages/carbon-web-components/examples/codesandbox/data-table/sandbox.config.json
@@ -1,3 +1,0 @@
-{
-  "template": "node"
-}
