+++ packages/carbon-web-components/src/components/data-table/stories/data-table-sorting-story.ts
@@ -9,13 +9,11 @@
 
 import { html } from 'lit';
 import { boolean, select, text } from '@storybook/addon-knobs';
-import { prefix } from '../../../../globals/settings';
-import { TABLE_SIZE } from '../../table';
-// @ts-ignore
-import Download16 from '@carbon/web-components/es/icons/download/16';
+import { prefix } from '../../../globals/settings';
+import { TABLE_SIZE } from '../table';
 // @ts-ignore
 import Settings16 from '@carbon/web-components/es/icons/settings/16';
-import '../../index';
+import '../index';
 import storyDocs from './data-table-story.mdx';
 
 const sizes = {
@@ -28,10 +26,10 @@ const sizes = {
 
 export const Default = () => {
   return html`
-    <cds-table expandable>
+    <cds-table is-sortable>
       <cds-table-header-title slot="title">DataTable</cds-table-header-title>
       <cds-table-header-description slot="description"
-        >With expansion</cds-table-header-description
+        >With filtering</cds-table-header-description
       >
 
       <cds-table-head>
@@ -55,10 +53,6 @@ export const Default = () => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 1</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -67,10 +61,6 @@ export const Default = () => {
           <cds-table-cell>Maureen's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 2</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -79,10 +69,6 @@ export const Default = () => {
           <cds-table-cell>Andrew's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 6</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -93,10 +79,6 @@ export const Default = () => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 4</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -105,10 +87,6 @@ export const Default = () => {
           <cds-table-cell>Mel's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 5</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -117,110 +95,6 @@ export const Default = () => {
           <cds-table-cell>Ronja's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-      </cds-table-body>
-    </cds-table>
-  `;
-};
-
-export const BatchExpansion = () => {
-  return html`
-    <cds-table expandable batch-expansion>
-      <cds-table-header-title slot="title">DataTable</cds-table-header-title>
-      <cds-table-header-description slot="description"
-        >With expansion</cds-table-header-description
-      >
-
-      <cds-table-head>
-        <cds-table-header-row>
-          <cds-table-header-cell>Name</cds-table-header-cell>
-          <cds-table-header-cell>Protocol</cds-table-header-cell>
-          <cds-table-header-cell>Port</cds-table-header-cell>
-          <cds-table-header-cell>Rule</cds-table-header-cell>
-          <cds-table-header-cell>Attached groups</cds-table-header-cell>
-          <cds-table-header-cell>Status</cds-table-header-cell>
-        </cds-table-header-row>
-      </cds-table-head>
-      <cds-table-body>
-        <cds-table-row>
-          <cds-table-cell>Load Balancer 3</cds-table-cell>
-          <cds-table-cell>HTTP</cds-table-cell>
-          <cds-table-cell>3000</cds-table-cell>
-          <cds-table-cell>Round robin</cds-table-cell>
-          <cds-table-cell>Kevin's VM Groups</cds-table-cell>
-          <cds-table-cell
-            ><cds-link disabled>Disabled</cds-link></cds-table-cell
-          >
-        </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row>
-          <cds-table-cell>Load Balancer 1</cds-table-cell>
-          <cds-table-cell>HTTP</cds-table-cell>
-          <cds-table-cell>443</cds-table-cell>
-          <cds-table-cell>Round robin</cds-table-cell>
-          <cds-table-cell>Maureen's VM Groups</cds-table-cell>
-          <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
-        </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row>
-          <cds-table-cell>Load Balancer 2</cds-table-cell>
-          <cds-table-cell>HTTP</cds-table-cell>
-          <cds-table-cell>80</cds-table-cell>
-          <cds-table-cell>DNS delegation</cds-table-cell>
-          <cds-table-cell>Andrew's VM Groups</cds-table-cell>
-          <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
-        </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row>
-          <cds-table-cell>Load Balancer 6</cds-table-cell>
-          <cds-table-cell>HTTP</cds-table-cell>
-          <cds-table-cell>3000</cds-table-cell>
-          <cds-table-cell>Round robin</cds-table-cell>
-          <cds-table-cell>Marc's VM Groups</cds-table-cell>
-          <cds-table-cell
-            ><cds-link disabled>Disabled</cds-link></cds-table-cell
-          >
-        </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row>
-          <cds-table-cell>Load Balancer 4</cds-table-cell>
-          <cds-table-cell>HTTP</cds-table-cell>
-          <cds-table-cell>443</cds-table-cell>
-          <cds-table-cell>Round robin</cds-table-cell>
-          <cds-table-cell>Mel's VM Groups</cds-table-cell>
-          <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
-        </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row>
-          <cds-table-cell>Load Balancer 5</cds-table-cell>
-          <cds-table-cell>HTTP</cds-table-cell>
-          <cds-table-cell>80</cds-table-cell>
-          <cds-table-cell>DNS delegation</cds-table-cell>
-          <cds-table-cell>Ronja's VM Groups</cds-table-cell>
-          <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
-        </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
       </cds-table-body>
     </cds-table>
   `;
@@ -231,7 +105,6 @@ export const Playground = (args) => {
     args?.[`${prefix}-table`] ?? {};
   return html`
     <cds-table
-      expandable
       ?is-sortable=${isSortable}
       locale="${locale}"
       size="${size}"
@@ -239,7 +112,7 @@ export const Playground = (args) => {
       ?use-zebra-styles="${useZebraStyles}">
       <cds-table-header-title slot="title">DataTable</cds-table-header-title>
       <cds-table-header-description slot="description"
-        >With expansion</cds-table-header-description
+        >With filtering</cds-table-header-description
       >
 
       <cds-table-head>
@@ -263,10 +136,6 @@ export const Playground = (args) => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 1</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -275,10 +144,6 @@ export const Playground = (args) => {
           <cds-table-cell>Maureen's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 2</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -287,10 +152,6 @@ export const Playground = (args) => {
           <cds-table-cell>Andrew's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 6</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -301,10 +162,6 @@ export const Playground = (args) => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 4</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -313,10 +170,6 @@ export const Playground = (args) => {
           <cds-table-cell>Mel's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
         <cds-table-row>
           <cds-table-cell>Load Balancer 5</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
@@ -325,10 +178,6 @@ export const Playground = (args) => {
           <cds-table-cell>Ronja's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
       </cds-table-body>
     </cds-table>
   `;
@@ -337,7 +186,7 @@ export const Playground = (args) => {
 Playground.parameters = {
   knobs: {
     [`${prefix}-table`]: () => ({
-      isSortable: boolean('Is sortable', false),
+      isSortable: boolean('Is sortable', true),
       locale: text('Locale', 'en'),
       size: select('Size', sizes, TABLE_SIZE.LG),
       useStaticWidth: boolean('Use static width', false),
@@ -347,7 +196,7 @@ Playground.parameters = {
 };
 
 export default {
-  title: 'Components/DataTable/Expansion',
+  title: 'Components/DataTable/Sorting',
   parameters: {
     ...storyDocs.parameters,
   },
