+++ packages/carbon-web-components/src/components/data-table/stories/data-table-story.mdx
@@ -1,5 +1,5 @@
 import { Props, Description } from '@storybook/addon-docs/blocks';
-import { cdnJs, cdnCss } from '../../../../globals/internal/storybook-cdn';
+import { cdnJs, cdnCss } from '../../../globals/internal/storybook-cdn';
 
 # Data table
